backend_server/requirements.txt
fastapi
uvicorn
python-jose
websockets
pydantic
loguru